CB: add comments to src code and misc enhancements

+++ src/dollar/style/width.js
@@ -1,4 +1,11 @@
+/**
+ * Get the current width of the first element in the set
+ * This method does not set values. Use .css() instead.
+ * @module style
+ * @returns Width
+ * @example $('div').width()
+ */
 
 $.fn.width = function () {
-    return parseInt(this.eq(0).css('width')) || 0;
+    return parseFloat(this.eq(0).css('width')) || 0;
 };
